@@ -2,6 +2,8 @@
 
 namespace Database\Seeders;
 
+use App\Models\Kategori;
+use App\Models\User;
 // use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
@@ -12,11 +14,41 @@ class DatabaseSeeder extends Seeder
      */
     public function run(): void
     {
-        // \App\Models\User::factory(10)->create();
+        // User::factory(10)->create();
 
-        // \App\Models\User::factory()->create([
-        //     'name' => 'Test User',
-        //     'email' => 'test@example.com',
-        // ]);
+        // USER
+        User::factory()->create([
+            'nama' => 'Administrator',
+            'email' => 'admin@gmail.com',
+            'role' => '1',
+            'status' => 1,
+            'hp' => '08123456789',
+            'password' => bcrypt('P@55word'),
+        ]);
+        User::factory()->create([
+            'nama' => 'Mohammad Ichwan Al Ghifari',
+            'email' => 'Ichwan@gmail.com',
+            'role' => '0',
+            'status' => 1,
+            'hp' => '08234567890',
+            'password' => bcrypt('P@55word'),
+        ]);
+
+        // KATEGORI
+        Kategori::create([
+            'nama_kategori' => 'Monitor'
+        ]);
+        Kategori::create([
+            'nama_kategori' => 'Keyboard'
+        ]);
+        Kategori::create([
+            'nama_kategori' => 'Flashdisk'
+        ]);
+        Kategori::create([
+            'nama_kategori' => 'Mouse'
+        ]);
+        Kategori::create([
+            'nama_kategori' => 'Cooling Pad'
+        ]);
     }
 }
